Renamed variables to reference ayet instead of freecash. Helps with consistency.
