+++ Functions/ayet_info.py
